+++ libs/e2e/src/lib/fixtures/orgs.ts
@@ -9,7 +9,7 @@ export enum ORG {
 
 export class Orgs {
   get(query: QueryInferface) : Partial<Organization>[] {
-    const orgSet: Partial<Organization>[] = 
+    const orgSet: Partial<Organization>[] =
                   (query.exist) ? orgsFixture : null;
 
     if (query.index && query.index !== -1) {
@@ -32,3 +32,35 @@ export class Orgs {
   }
 
 }
+
+export const ORGANIZATION: Organization = {
+  id: 'Cy1234',
+  email: `dev+${Date.now()}@cascade8.com`,
+  address: {
+    street: '42 test road',
+    zipCode: '69001',
+    city: 'Testville',
+    country: 'France',
+    phoneNumber: '+334 857 953'
+  },
+  activity: 'Distribution',
+  fiscalNumber: '95 14 958 641 215 C',
+
+  bankAccount: {
+    address: {
+      street: '21 gold street',
+      zipCode: '69001',
+      city: 'Moneytown',
+      country: 'Germany'
+    },
+    IBAN: 'FR1420041010050500013M02606',
+    BIC: 'CCBPFRPPVER',
+    bankName: 'Cypress Bank',
+    holderName: 'Cypress'
+  },
+
+  denomination: {
+    full: `Cypress & Party - ${Date.now()}`,
+    public: 'Cypress & Party'
+  }
+};
